TST: collect Interval tests (#56072)
* TST: collect scalar Interval tests

* TST: collect IntervalIndex tests

* TST: collect Interval tests
