+++ pandas/tests/scalar/interval/test_interval.py
@@ -3,12 +3,9 @@ import pytest
 
 from pandas import (
     Interval,
-    Period,
     Timedelta,
     Timestamp,
 )
-import pandas._testing as tm
-import pandas.core.common as com
 
 
 @pytest.fixture
@@ -22,48 +19,6 @@ class TestInterval:
         assert interval.left == 0
         assert interval.right == 1
         assert interval.mid == 0.5
-
-    def test_repr(self, interval):
-        assert repr(interval) == "Interval(0, 1, closed='right')"
-        assert str(interval) == "(0, 1]"
-
-        interval_left = Interval(0, 1, closed="left")
-        assert repr(interval_left) == "Interval(0, 1, closed='left')"
-        assert str(interval_left) == "[0, 1)"
-
-    def test_contains(self, interval):
-        assert 0.5 in interval
-        assert 1 in interval
-        assert 0 not in interval
-
-        interval_both = Interval(0, 1, "both")
-        assert 0 in interval_both
-        assert 1 in interval_both
-
-        interval_neither = Interval(0, 1, closed="neither")
-        assert 0 not in interval_neither
-        assert 0.5 in interval_neither
-        assert 1 not in interval_neither
-
-    def test_equal(self):
-        assert Interval(0, 1) == Interval(0, 1, closed="right")
-        assert Interval(0, 1) != Interval(0, 1, closed="left")
-        assert Interval(0, 1) != 0
-
-    def test_comparison(self):
-        msg = (
-            "'<' not supported between instances of "
-            "'pandas._libs.interval.Interval' and 'int'"
-        )
-        with pytest.raises(TypeError, match=msg):
-            Interval(0, 1) < 2
-
-        assert Interval(0, 1) < Interval(1, 2)
-        assert Interval(0, 1) < Interval(0, 2)
-        assert Interval(0, 1) < Interval(0.5, 1.5)
-        assert Interval(0, 1) <= Interval(0, 1)
-        assert Interval(0, 1) > Interval(-1, 2)
-        assert Interval(0, 1) >= Interval(0, 1)
 
     def test_hash(self, interval):
         # should not raise
@@ -130,150 +85,3 @@ class TestInterval:
         result = iv.is_empty
         expected = closed != "both"
         assert result is expected
-
-    @pytest.mark.parametrize(
-        "left, right",
-        [
-            ("a", "z"),
-            (("a", "b"), ("c", "d")),
-            (list("AB"), list("ab")),
-            (Interval(0, 1), Interval(1, 2)),
-            (Period("2018Q1", freq="Q"), Period("2018Q1", freq="Q")),
-        ],
-    )
-    def test_construct_errors(self, left, right):
-        # GH 23013
-        msg = "Only numeric, Timestamp and Timedelta endpoints are allowed"
-        with pytest.raises(ValueError, match=msg):
-            Interval(left, right)
-
-    def test_math_add(self, closed):
-        interval = Interval(0, 1, closed=closed)
-        expected = Interval(1, 2, closed=closed)
-
-        result = interval + 1
-        assert result == expected
-
-        result = 1 + interval
-        assert result == expected
-
-        result = interval
-        result += 1
-        assert result == expected
-
-        msg = r"unsupported operand type\(s\) for \+"
-        with pytest.raises(TypeError, match=msg):
-            interval + interval
-
-        with pytest.raises(TypeError, match=msg):
-            interval + "foo"
-
-    def test_math_sub(self, closed):
-        interval = Interval(0, 1, closed=closed)
-        expected = Interval(-1, 0, closed=closed)
-
-        result = interval - 1
-        assert result == expected
-
-        result = interval
-        result -= 1
-        assert result == expected
-
-        msg = r"unsupported operand type\(s\) for -"
-        with pytest.raises(TypeError, match=msg):
-            interval - interval
-
-        with pytest.raises(TypeError, match=msg):
-            interval - "foo"
-
-    def test_math_mult(self, closed):
-        interval = Interval(0, 1, closed=closed)
-        expected = Interval(0, 2, closed=closed)
-
-        result = interval * 2
-        assert result == expected
-
-        result = 2 * interval
-        assert result == expected
-
-        result = interval
-        result *= 2
-        assert result == expected
-
-        msg = r"unsupported operand type\(s\) for \*"
-        with pytest.raises(TypeError, match=msg):
-            interval * interval
-
-        msg = r"can\'t multiply sequence by non-int"
-        with pytest.raises(TypeError, match=msg):
-            interval * "foo"
-
-    def test_math_div(self, closed):
-        interval = Interval(0, 1, closed=closed)
-        expected = Interval(0, 0.5, closed=closed)
-
-        result = interval / 2.0
-        assert result == expected
-
-        result = interval
-        result /= 2.0
-        assert result == expected
-
-        msg = r"unsupported operand type\(s\) for /"
-        with pytest.raises(TypeError, match=msg):
-            interval / interval
-
-        with pytest.raises(TypeError, match=msg):
-            interval / "foo"
-
-    def test_math_floordiv(self, closed):
-        interval = Interval(1, 2, closed=closed)
-        expected = Interval(0, 1, closed=closed)
-
-        result = interval // 2
-        assert result == expected
-
-        result = interval
-        result //= 2
-        assert result == expected
-
-        msg = r"unsupported operand type\(s\) for //"
-        with pytest.raises(TypeError, match=msg):
-            interval // interval
-
-        with pytest.raises(TypeError, match=msg):
-            interval // "foo"
-
-    def test_constructor_errors(self):
-        msg = "invalid option for 'closed': foo"
-        with pytest.raises(ValueError, match=msg):
-            Interval(0, 1, closed="foo")
-
-        msg = "left side of interval must be <= right side"
-        with pytest.raises(ValueError, match=msg):
-            Interval(1, 0)
-
-    @pytest.mark.parametrize(
-        "tz_left, tz_right", [(None, "UTC"), ("UTC", None), ("UTC", "US/Eastern")]
-    )
-    def test_constructor_errors_tz(self, tz_left, tz_right):
-        # GH 18538
-        left = Timestamp("2017-01-01", tz=tz_left)
-        right = Timestamp("2017-01-02", tz=tz_right)
-
-        if com.any_none(tz_left, tz_right):
-            error = TypeError
-            msg = "Cannot compare tz-naive and tz-aware timestamps"
-        else:
-            error = ValueError
-            msg = "left and right must have the same time zone"
-        with pytest.raises(error, match=msg):
-            Interval(left, right)
-
-    def test_equality_comparison_broadcasts_over_array(self):
-        # https://github.com/pandas-dev/pandas/issues/35931
-        interval = Interval(0, 1)
-        arr = np.array([interval, interval])
-        result = interval == arr
-        expected = np.array([True, True])
-        tm.assert_numpy_array_equal(result, expected)
